Type hint the team
