+++ tests/Models/Team.php
@@ -2,10 +2,10 @@
 
 namespace Laratrust\Tests\Models;
 
-use Laratrust\Models\LaratrustTeam;
+use Laratrust\Models\Team as TeamModel;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Team extends LaratrustTeam
+class Team extends TeamModel
 {
     use SoftDeletes;
